Migrated DAL for FHIR Servers
