+++ query-connector/src/app/(pages)/userManagement/components/RoleDropdown.tsx
@@ -1,7 +1,7 @@
 import { Label } from "@trussworks/react-uswds";
 import classNames from "classnames";
 import React from "react";
-import { RoleTypeValues } from "../../../models/entities/user-management";
+import { RoleTypeValues } from "../../../models/entities/users";
 
 export interface RoleDropdownProps {
   id: string;
