+++ query-connector/src/app/models/entities/users.ts
