+++ query-connector/src/app/ui/components/withAuth/WithAuth.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useContext } from "react";
-import { RoleTypeValues } from "@/app/models/entities/user-management";
+import { RoleTypeValues } from "@/app/models/entities/users";
 import { useSession } from "next-auth/react";
 import { redirect, usePathname } from "next/navigation";
 import { pagesConfig } from "@/app/shared/page-routes";
